First try at in-place (it doesn't work though)
